comments
Thanks for adding in this _info file for me to see what a sqlite table looks like! I have never seen one and this gives me a great idea of how it functions on your site and where the data comes from. I had no idea a picture was referred to as a BLOB in sql queries.  Also, love the addition of Harry Potter movies, I'd shop at this e-commerce site 👍