Update register.ejs
I love your landing page and the use of the book image. It's very simple and clean and looks great. 
One suggestion I have for the register page is adding a href to link back to the login page in case it was reached by mistake. Similar to how you have the link for "Don't have an account? Register now!" on the login page. A lot of websites these days have an additional link on the register page that says "Already have an account? Log in here". It's not needed but I think it could be a great addition to eliminate confusion for visitors to your site!